+++ webpush/store_postgres.go
@@ -4,7 +4,7 @@ import (
 	"database/sql"
 	"fmt"
 
-	ntfydb "heckel.io/ntfy/v2/db"
+	"heckel.io/ntfy/v2/db"
 )
 
 const (
@@ -73,7 +73,7 @@ const (
 )
 
 // NewPostgresStore creates a new PostgreSQL-backed web push store using an existing database connection pool.
-func NewPostgresStore(d *ntfydb.DB) (*Store, error) {
+func NewPostgresStore(d *db.DB) (*Store, error) {
 	if err := setupPostgres(d.Primary()); err != nil {
 		return nil, err
 	}
@@ -97,11 +97,11 @@ func NewPostgresStore(d *ntfydb.DB) (*Store, error) {
 	}, nil
 }
 
-func setupPostgres(db *sql.DB) error {
+func setupPostgres(d *sql.DB) error {
 	var schemaVersion int
-	err := db.QueryRow(postgresSelectSchemaVersionQuery).Scan(&schemaVersion)
+	err := d.QueryRow(postgresSelectSchemaVersionQuery).Scan(&schemaVersion)
 	if err != nil {
-		return setupNewPostgres(db)
+		return setupNewPostgres(d)
 	}
 	if schemaVersion > pgCurrentSchemaVersion {
 		return fmt.Errorf("unexpected schema version: version %d is higher than current version %d", schemaVersion, pgCurrentSchemaVersion)
@@ -109,8 +109,8 @@ func setupPostgres(db *sql.DB) error {
 	return nil
 }
 
-func setupNewPostgres(sqlDB *sql.DB) error {
-	return ntfydb.ExecTx(sqlDB, func(tx *sql.Tx) error {
+func setupNewPostgres(d *sql.DB) error {
+	return db.ExecTx(d, func(tx *sql.Tx) error {
 		if _, err := tx.Exec(postgresCreateTablesQuery); err != nil {
 			return err
 		}
